video-to-svg/requirements.txt
opencv-python
numpy
svgwrite
PyYAML